cont validations functions

+++ libftx/include/ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mkitano <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: mkitano <mkitano@student.42sp.org.br>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/13 18:31:44 by mkitano           #+#    #+#             */
-/*   Updated: 2025/08/13 18:31:47 by mkitano          ###   ########.fr       */
+/*   Updated: 2025/08/15 17:19:00 by mkitano          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int		ft_putchar(char c);
 int		ft_putstr(char *str);
 int		ft_putnbr_dec(int nbr);
 int		ft_putnbr_u(unsigned int nbr);
-int		ft_putnbr_hex(unsigned int nbr, char specifier);
+int		ft_putnbr_hex(unsigned long nbr, char specifier);
 int		ft_address_hex(void *ptr);
 
 int		ft_printf(const char *format, ...);
